Split ProjectiveDecomposition into two classes.
Class DiagonalScaling expresses a common mathematical paradigm for balancing matrices.
Class ScalingDecomposition extends DiagonalScaling to express both existing and new methods of matrix balancing or matrix normalization, similar operations done for different reasons. Matrix balancing is not usually presented as a type of matrix decomposition, but it is equivalent to a matrix decomposition that only involves scaling. Projective decomposition is a type of scaling decomposition that I have introduced for normalization purposes.